Final v7.4.1

+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/GamesInProgress.java
@@ -172,7 +172,7 @@ public class GamesInProgress {
 		public boolean daily;
 		public boolean dailyReplay;
 
-		public int level;
+		public long level;
 		public int str;
 		public int strBonus;
 		public long exp;
